PA1_Bergdahl::parse
Module parse

Structs

Params

A configuration struct to specify params used in the Global and Local Alignment Algorithms.

Functions

extract_sequences

get_config

Extracts information from the file with the inputted name into a Param struct instance.

split_lines 🔒